refactor: move tools to core
