+++ src/core/tools/ruler.ts
